+++ tools/qa/src/test/java/com/tonikelope/coronapoker/BlindStructureTest.java
@@ -37,10 +37,10 @@ import org.junit.jupiter.api.Test;
  */
 public class BlindStructureTest {
 
-    private static float[][] levels(float... sbBbPairs) {
-        float[][] out = new float[sbBbPairs.length / 2][];
+    private static double[][] levels(double... sbBbPairs) {
+        double[][] out = new double[sbBbPairs.length / 2][];
         for (int i = 0; i < out.length; i++) {
-            out[i] = new float[]{sbBbPairs[2 * i], sbBbPairs[2 * i + 1]};
+            out[i] = new double[]{sbBbPairs[2 * i], sbBbPairs[2 * i + 1]};
         }
         return out;
     }
@@ -50,16 +50,16 @@ public class BlindStructureTest {
     @Test
     void defaultLadderHasTwentyLevelsAllDoubleBlind() {
         // Arrange / Act
-        float[][] def = BlindStructure.defaultLevels();
+        double[][] def = BlindStructure.defaultLevels();
 
         // Assert: matches the twenty hard-coded combo entries, bb = 2*sb.
         assertEquals(20, def.length);
-        assertArrayEquals(new float[]{0.1f, 0.2f}, def[0], 0f);
-        assertArrayEquals(new float[]{0.5f, 1f}, def[3], 0f);
-        assertArrayEquals(new float[]{1f, 2f}, def[4], 0f);
-        assertArrayEquals(new float[]{5000f, 10000f}, def[19], 0f);
-        for (float[] lvl : def) {
-            assertEquals(lvl[0] * 2f, lvl[1], 0f);
+        assertArrayEquals(new double[]{0.1, 0.2}, def[0], 0);
+        assertArrayEquals(new double[]{0.5, 1}, def[3], 0);
+        assertArrayEquals(new double[]{1, 2}, def[4], 0);
+        assertArrayEquals(new double[]{5000, 10000}, def[19], 0);
+        for (double[] lvl : def) {
+            assertEquals(lvl[0] * 2, lvl[1], 0);
         }
     }
 
@@ -72,9 +72,9 @@ public class BlindStructureTest {
 
     @Test
     void defaultLadderReturnsAFreshArrayEachCall() {
-        float[][] a = BlindStructure.defaultLevels();
-        a[0][0] = 999f;
-        assertEquals(0.1f, BlindStructure.defaultLevels()[0][0], 0f);
+        double[][] a = BlindStructure.defaultLevels();
+        a[0][0] = 999;
+        assertEquals(0.1, BlindStructure.defaultLevels()[0][0], 0);
     }
 
     // ----- Serialization round-trip -------------------------------------------
@@ -82,11 +82,11 @@ public class BlindStructureTest {
     @Test
     void levelsStringRoundTripsWholeAndFractionalValues() {
         // Arrange: mix of fractional and whole blinds, non-2x big blind.
-        float[][] in = levels(0.5f, 1f, 1f, 2f, 25f, 50f, 100f, 250f);
+        double[][] in = levels(0.5, 1, 1, 2, 25, 50, 100, 250);
 
         // Act
         String csv = BlindStructure.levelsToString(in);
-        float[][] out = BlindStructure.parseLevels(csv);
+        double[][] out = BlindStructure.parseLevels(csv);
 
         // Assert: compact form, exact round-trip.
         assertEquals("0.5/1,1/2,25/50,100/250", csv);
@@ -95,31 +95,31 @@ public class BlindStructureTest {
 
     @Test
     void formatLevelValueNeverAbbreviatesLargeBlinds() {
-        // B1 regression guard: Helpers.float2String renders >=1000 as "1K"/"5K",
+        // B1 regression guard: Helpers.double2String renders >=1000 as "1K"/"5K",
         // which the Float.valueOf combo parsers cannot read (NumberFormatException
         // on create / edit-blinds / UPDATEBLINDS broadcast). formatLevelValue must
         // emit a plain, parseable number for any magnitude.
-        for (float v : new float[]{1000f, 2000f, 5000f, 10000f, 3000f}) {
+        for (double v : new double[]{1000, 2000, 5000, 10000, 3000}) {
             String s = BlindStructure.formatLevelValue(v);
             assertFalse(s.toLowerCase().contains("k"), s);
-            assertEquals(v, Float.parseFloat(s.replace(",", ".")), 0f);
+            assertEquals(v, Float.parseFloat(s.replace(",", ".")), 0);
         }
         // A full tournament level round-trips through the "small / big" form.
-        String lvl = BlindStructure.formatLevel(1000f, 2000f);
+        String lvl = BlindStructure.formatLevel(1000, 2000);
         assertEquals("1000 / 2000", lvl);
         String[] parts = lvl.replace(",", ".").split("/");
-        assertEquals(1000f, Float.parseFloat(parts[0].trim()), 0f);
-        assertEquals(2000f, Float.parseFloat(parts[1].trim()), 0f);
+        assertEquals(1000, Float.parseFloat(parts[0].trim()), 0);
+        assertEquals(2000, Float.parseFloat(parts[1].trim()), 0);
     }
 
     @Test
     void levelsStringIsSafeForTheInitWire() {
         // The INIT command is '#'-delimited with '@' sub-delimiters; the serialized
         // ladder must never contain either or it would corrupt the wire framing.
-        for (float[][] ladder : new float[][][]{
+        for (double[][] ladder : new double[][][]{
             BlindStructure.defaultLevels(),
-            levels(0.1f, 0.2f, 0.5f, 1f),
-            levels(10f, 25f, 20f, 50f, 40f, 100f)}) {
+            levels(0.1, 0.2, 0.5, 1),
+            levels(10, 25, 20, 50, 40, 100)}) {
             String csv = BlindStructure.levelsToString(ladder);
             assertFalse(csv.contains("#"), csv);
             assertFalse(csv.contains("@"), csv);
@@ -129,15 +129,15 @@ public class BlindStructureTest {
 
     @Test
     void levelsStringRoundTripsNickelBlinds() {
-        float[][] in = levels(0.05f, 0.10f, 0.25f, 0.50f, 0.35f, 0.70f);
+        double[][] in = levels(0.05, 0.10, 0.25, 0.50, 0.35, 0.70);
         String csv = BlindStructure.levelsToString(in);
         assertArrayEquals(in, BlindStructure.parseLevels(csv));
     }
 
     @Test
     void parseLevelsToleratesWhitespace() {
-        float[][] out = BlindStructure.parseLevels(" 25/50 , 50/100 ");
-        assertArrayEquals(levels(25f, 50f, 50f, 100f), out);
+        double[][] out = BlindStructure.parseLevels(" 25/50 , 50/100 ");
+        assertArrayEquals(levels(25, 50, 50, 100), out);
     }
 
     @Test
@@ -151,7 +151,7 @@ public class BlindStructureTest {
     @Test
     void parseValidatedLevelsAcceptsAValidLadder() {
         // Grammar-valid AND logically valid (incl. a non-2x level) round-trips.
-        assertArrayEquals(levels(0.10f, 0.25f, 0.20f, 0.50f),
+        assertArrayEquals(levels(0.10, 0.25, 0.20, 0.50),
                 BlindStructure.parseValidatedLevels("0.1/0.25,0.2/0.5"));
     }
 
@@ -193,22 +193,22 @@ public class BlindStructureTest {
 
     @Test
     void levelsValidationAcceptsSaneLadders() {
-        assertNull(BlindStructure.validateLevels(levels(25f, 50f, 50f, 100f, 100f, 200f)));
+        assertNull(BlindStructure.validateLevels(levels(25, 50, 50, 100, 100, 200)));
         // Single fixed level is allowed (escalation simply never moves).
-        assertNull(BlindStructure.validateLevels(levels(100f, 200f)));
+        assertNull(BlindStructure.validateLevels(levels(100, 200)));
         // Non-2x big blind is allowed (total-freedom mode).
-        assertNull(BlindStructure.validateLevels(levels(10f, 25f, 20f, 50f)));
+        assertNull(BlindStructure.validateLevels(levels(10, 25, 20, 50)));
         // Big blind equal to small blind is allowed (bb >= sb).
-        assertNull(BlindStructure.validateLevels(levels(10f, 10f, 20f, 20f)));
+        assertNull(BlindStructure.validateLevels(levels(10, 10, 20, 20)));
     }
 
     @Test
     void levelsValidationRejectsEmptyAndOversized() {
         assertEquals(BlindStructure.ERR_NO_LEVELS, BlindStructure.validateLevels(null));
-        assertEquals(BlindStructure.ERR_NO_LEVELS, BlindStructure.validateLevels(new float[0][]));
-        float[][] tooMany = new float[BlindStructure.MAX_LEVELS + 1][];
+        assertEquals(BlindStructure.ERR_NO_LEVELS, BlindStructure.validateLevels(new double[0][]));
+        double[][] tooMany = new double[BlindStructure.MAX_LEVELS + 1][];
         for (int i = 0; i < tooMany.length; i++) {
-            tooMany[i] = new float[]{i + 1, (i + 1) * 2};
+            tooMany[i] = new double[]{i + 1, (i + 1) * 2};
         }
         assertEquals(BlindStructure.ERR_TOO_MANY_LEVELS, BlindStructure.validateLevels(tooMany));
     }
@@ -216,53 +216,53 @@ public class BlindStructureTest {
     @Test
     void levelsValidationAcceptsNickelGranularBlinds() {
         // Blinds in 0.05 steps, minimum 0.05 (0.25/0.35 allowed, 0.33 not).
-        assertNull(BlindStructure.validateLevels(levels(0.05f, 0.10f)));
-        assertNull(BlindStructure.validateLevels(levels(0.25f, 0.50f, 0.50f, 1.00f)));
-        assertNull(BlindStructure.validateLevels(levels(0.35f, 0.70f)));
-        assertNull(BlindStructure.validateLevels(levels(0.15f, 0.30f)));
+        assertNull(BlindStructure.validateLevels(levels(0.05, 0.10)));
+        assertNull(BlindStructure.validateLevels(levels(0.25, 0.50, 0.50, 1.00)));
+        assertNull(BlindStructure.validateLevels(levels(0.35, 0.70)));
+        assertNull(BlindStructure.validateLevels(levels(0.15, 0.30)));
     }
 
     @Test
     void levelsValidationRejectsOutOfRangeAndPrecision() {
-        assertEquals(BlindStructure.ERR_VALUE_RANGE, BlindStructure.validateLevels(levels(0f, 0f)));
+        assertEquals(BlindStructure.ERR_VALUE_RANGE, BlindStructure.validateLevels(levels(0, 0)));
         assertEquals(BlindStructure.ERR_VALUE_RANGE,
-                BlindStructure.validateLevels(levels(0.04f, 0.10f))); // below MIN_BLIND (0.05)
+                BlindStructure.validateLevels(levels(0.04, 0.10))); // below MIN_BLIND (0.05)
         assertEquals(BlindStructure.ERR_PRECISION,
-                BlindStructure.validateLevels(levels(0.33f, 0.66f))); // not a 0.05 multiple
+                BlindStructure.validateLevels(levels(0.33, 0.66))); // not a 0.05 multiple
         assertEquals(BlindStructure.ERR_PRECISION,
-                BlindStructure.validateLevels(levels(0.07f, 0.20f))); // not a 0.05 multiple
+                BlindStructure.validateLevels(levels(0.07, 0.20))); // not a 0.05 multiple
     }
 
     @Test
     void levelsValidationRejectsBbBelowSb() {
-        assertEquals(BlindStructure.ERR_BB_LT_SB, BlindStructure.validateLevels(levels(50f, 25f)));
+        assertEquals(BlindStructure.ERR_BB_LT_SB, BlindStructure.validateLevels(levels(50, 25)));
     }
 
     @Test
     void levelsValidationRejectsNonIncreasingLadder() {
         // Small blind not climbing.
         assertEquals(BlindStructure.ERR_NOT_INCREASING,
-                BlindStructure.validateLevels(levels(50f, 100f, 50f, 200f)));
+                BlindStructure.validateLevels(levels(50, 100, 50, 200)));
         // Big blind not climbing.
         assertEquals(BlindStructure.ERR_NOT_INCREASING,
-                BlindStructure.validateLevels(levels(50f, 100f, 60f, 100f)));
+                BlindStructure.validateLevels(levels(50, 100, 60, 100)));
         // Duplicate level.
         assertEquals(BlindStructure.ERR_NOT_INCREASING,
-                BlindStructure.validateLevels(levels(50f, 100f, 50f, 100f)));
+                BlindStructure.validateLevels(levels(50, 100, 50, 100)));
     }
 
     @Test
     void constructorThrowsOnInvalidInput() {
-        assertThrows(IllegalArgumentException.class, () -> new BlindStructure("", levels(25f, 50f)));
-        assertThrows(IllegalArgumentException.class, () -> new BlindStructure("ok", levels(50f, 25f)));
+        assertThrows(IllegalArgumentException.class, () -> new BlindStructure("", levels(25, 50)));
+        assertThrows(IllegalArgumentException.class, () -> new BlindStructure("ok", levels(50, 25)));
     }
 
     @Test
     void getLevelsIsADefensiveCopy() {
-        BlindStructure bs = new BlindStructure("Casa", levels(25f, 50f, 50f, 100f));
-        float[][] got = bs.getLevels();
-        got[0][0] = 999f;
-        assertEquals(25f, bs.getLevels()[0][0], 0f);
+        BlindStructure bs = new BlindStructure("Casa", levels(25, 50, 50, 100));
+        double[][] got = bs.getLevels();
+        got[0][0] = 999;
+        assertEquals(25, bs.getLevels()[0][0], 0);
     }
 
     // ----- Registry persistence (pure, over a Properties object) --------------
@@ -271,8 +271,8 @@ public class BlindStructureTest {
     void registryRoundTripPreservesOrderAndValues() {
         // Arrange
         List<BlindStructure> in = new ArrayList<>();
-        in.add(new BlindStructure("Casa", levels(25f, 50f, 50f, 100f, 100f, 200f)));
-        in.add(new BlindStructure("Turbo", levels(50f, 100f, 100f, 200f)));
+        in.add(new BlindStructure("Casa", levels(25, 50, 50, 100, 100, 200)));
+        in.add(new BlindStructure("Turbo", levels(50, 100, 100, 200)));
         Properties props = new Properties();
 
         // Act
@@ -293,12 +293,12 @@ public class BlindStructureTest {
         Properties props = new Properties();
         props.setProperty("zoom_menu", "1.25");
         BlindStructure.writeTo(props, List.of(
-                new BlindStructure("Old1", levels(25f, 50f)),
-                new BlindStructure("Old2", levels(50f, 100f)),
-                new BlindStructure("Old3", levels(75f, 150f))));
+                new BlindStructure("Old1", levels(25, 50)),
+                new BlindStructure("Old2", levels(50, 100)),
+                new BlindStructure("Old3", levels(75, 150))));
 
         // Act: overwrite with a smaller set (a deletion).
-        BlindStructure.writeTo(props, List.of(new BlindStructure("Kept", levels(10f, 20f))));
+        BlindStructure.writeTo(props, List.of(new BlindStructure("Kept", levels(10, 20))));
         LinkedHashMap<String, BlindStructure> out = BlindStructure.readFrom(props);
 
         // Assert: only the new set survives; the deleted structures' keys are gone;
@@ -346,7 +346,7 @@ public class BlindStructureTest {
 
         assertEquals(1, out.size());
         // First wins.
-        assertArrayEquals(levels(25f, 50f), out.get("Casa").getLevels());
+        assertArrayEquals(levels(25, 50), out.get("Casa").getLevels());
     }
 
     @Test
@@ -375,49 +375,49 @@ public class BlindStructureTest {
     @Test
     void indexOfLevelFindsByCentResolution() {
         // Cent resolution distinguishes 0.25 from 0.20/0.30 (needed for 0.05 blinds).
-        float[][] s = levels(0.1f, 0.2f, 0.25f, 0.5f, 0.5f, 1f, 25f, 50f);
-        assertEquals(0, BlindStructure.indexOfLevel(s, 0.1f));
-        assertEquals(1, BlindStructure.indexOfLevel(s, 0.25f));
-        assertEquals(2, BlindStructure.indexOfLevel(s, 0.5f));
-        assertEquals(3, BlindStructure.indexOfLevel(s, 25f));
-        // Tolerant of float drift (rounds to the same cent).
-        assertEquals(1, BlindStructure.indexOfLevel(s, 0.250001f));
+        double[][] s = levels(0.1, 0.2, 0.25, 0.5, 0.5, 1, 25, 50);
+        assertEquals(0, BlindStructure.indexOfLevel(s, 0.1));
+        assertEquals(1, BlindStructure.indexOfLevel(s, 0.25));
+        assertEquals(2, BlindStructure.indexOfLevel(s, 0.5));
+        assertEquals(3, BlindStructure.indexOfLevel(s, 25));
+        // Tolerant of double drift (rounds to the same cent).
+        assertEquals(1, BlindStructure.indexOfLevel(s, 0.250001));
         // Off the ladder (0.27 is not a level).
-        assertEquals(-1, BlindStructure.indexOfLevel(s, 0.27f));
-        assertEquals(-1, BlindStructure.indexOfLevel(null, 1f));
+        assertEquals(-1, BlindStructure.indexOfLevel(s, 0.27));
+        assertEquals(-1, BlindStructure.indexOfLevel(null, 1));
     }
 
     @Test
     void nextLevelWalksThenCapsAtTop() {
-        float[][] s = levels(25f, 50f, 50f, 100f, 100f, 200f);
-        assertArrayEquals(new float[]{50f, 100f}, BlindStructure.nextLevel(s, 25f), 0f);
-        assertArrayEquals(new float[]{100f, 200f}, BlindStructure.nextLevel(s, 50f), 0f);
+        double[][] s = levels(25, 50, 50, 100, 100, 200);
+        assertArrayEquals(new double[]{50, 100}, BlindStructure.nextLevel(s, 25), 0);
+        assertArrayEquals(new double[]{100, 200}, BlindStructure.nextLevel(s, 50), 0);
         // Top level: no further escalation.
-        assertNull(BlindStructure.nextLevel(s, 100f));
+        assertNull(BlindStructure.nextLevel(s, 100));
         // Off the ladder.
-        assertNull(BlindStructure.nextLevel(s, 999f));
+        assertNull(BlindStructure.nextLevel(s, 999));
     }
 
     @Test
     void nextLevelDrivesAFullDeterministicWalk() {
         // Mirrors how Crupier.doblarCiegas climbs a custom structure: start at the
         // first level and follow nextLevel until it caps, collecting the ladder.
-        float[][] s = levels(10f, 25f, 20f, 50f, 40f, 100f); // non-2x big blinds
-        ArrayList<float[]> walked = new ArrayList<>();
+        double[][] s = levels(10, 25, 20, 50, 40, 100); // non-2x big blinds
+        ArrayList<double[]> walked = new ArrayList<>();
         walked.add(s[0]);
-        float sb = s[0][0];
-        float[] next;
+        double sb = s[0][0];
+        double[] next;
         while ((next = BlindStructure.nextLevel(s, sb)) != null) {
             walked.add(next);
             sb = next[0];
         }
-        assertArrayEquals(s, walked.toArray(new float[0][]));
+        assertArrayEquals(s, walked.toArray(new double[0][]));
     }
 
     @Test
     void singleLevelStructureNeverEscalates() {
-        float[][] s = levels(100f, 200f);
-        assertNull(BlindStructure.nextLevel(s, 100f));
+        double[][] s = levels(100, 200);
+        assertNull(BlindStructure.nextLevel(s, 100));
     }
 
     @Test
@@ -425,17 +425,17 @@ public class BlindStructureTest {
         // The helper caps the built-in ladder at 5000/10000; in-engine the default
         // (null) path keeps its infinite-by-decade behaviour, this only documents
         // the helper's contract for when a default-derived structure is walked.
-        float[][] def = BlindStructure.defaultLevels();
-        assertArrayEquals(new float[]{0.2f, 0.4f}, BlindStructure.nextLevel(def, 0.1f), 0f);
-        assertArrayEquals(new float[]{2f, 4f}, BlindStructure.nextLevel(def, 1f), 0f);
-        assertNull(BlindStructure.nextLevel(def, 5000f));
+        double[][] def = BlindStructure.defaultLevels();
+        assertArrayEquals(new double[]{0.2, 0.4}, BlindStructure.nextLevel(def, 0.1), 0);
+        assertArrayEquals(new double[]{2, 4}, BlindStructure.nextLevel(def, 1), 0);
+        assertNull(BlindStructure.nextLevel(def, 5000));
     }
 
     @Test
     void equalsAndHashCodeAreValueBased() {
-        BlindStructure a = new BlindStructure("Casa", levels(25f, 50f, 50f, 100f));
-        BlindStructure b = new BlindStructure("Casa", levels(25f, 50f, 50f, 100f));
-        BlindStructure c = new BlindStructure("Casa", levels(25f, 50f));
+        BlindStructure a = new BlindStructure("Casa", levels(25, 50, 50, 100));
+        BlindStructure b = new BlindStructure("Casa", levels(25, 50, 50, 100));
+        BlindStructure c = new BlindStructure("Casa", levels(25, 50));
         assertEquals(a, b);
         assertEquals(a.hashCode(), b.hashCode());
         assertFalse(a.equals(c));
